see previous commit message

+++ server/controllers/bookController.js
@@ -13,39 +13,54 @@ module.exports = {
     },
 
     createBook: (req, res) => {
-        const db = req.app.get('db');
-        const { book_title, book_subtitle, book_image, book_subject, author, membership_required_book, membership_ids_book, visible } = req.body
-        const title = book_title.toUpperCase();
-        const subject = book_subject.toUpperCase();
-        const authorFormat = `{${author}}`;
-        const idsFormat = `{${membership_ids_book}}`;
+        if(req.user.is_admin){
 
-        db.create_book([title, book_subtitle, book_image, subject, authorFormat, membership_required_book, membership_ids_book, visible]).then(newBook => {
-            res.status(200).send(newBook);
-        })
+            const db = req.app.get('db');
+            const { book_title, book_subtitle, book_image, book_subject, author, membership_required_book, membership_ids_book, visible } = req.body
+            const title = book_title.toUpperCase();
+            const subject = book_subject.toUpperCase();
+            const authorFormat = `{${author}}`;
+            const idsFormat = `{${membership_ids_book}}`;
+            
+            db.create_book([title, book_subtitle, book_image, subject, authorFormat, membership_required_book, membership_ids_book, visible]).then(newBook => {
+                res.status(200).send(newBook);
+            })
+        } else {
+            res.status(403).send('Unauthorized');
+        }
     },
 
     updateBook: (req, res) => {
-        const db = req.app.get('db');
-        const { bookId } = req.params
-        const { book_title, book_subtitle, book_image, book_subject, author, membership_required_book, membership_ids_book, visible } = req.body
-        const title = book_title.toUpperCase();
-        const subject = book_subject.toUpperCase();
-        const authorFormat = `{${author}}`
-        const idsFormat = `{${membership_ids_book}}`
-        
-        db.update_book([bookId, title, book_subtitle, book_image, subject, authorFormat, membership_required_book, membership_ids_book, visible]).then(udpatedBook => {
-            res.status(200).send(udpatedBook)
-        })
+        if(req.user.is_admin){
+
+            const db = req.app.get('db');
+            const { bookId } = req.params
+            const { book_title, book_subtitle, book_image, book_subject, author, membership_required_book, membership_ids_book, visible } = req.body
+            const title = book_title.toUpperCase();
+            const subject = book_subject.toUpperCase();
+            const authorFormat = `{${author}}`
+            const idsFormat = `{${membership_ids_book}}`
+            
+            db.update_book([bookId, title, book_subtitle, book_image, subject, authorFormat, membership_required_book, membership_ids_book, visible]).then(udpatedBook => {
+                res.status(200).send(udpatedBook)
+            })
+        } else {
+            res.status(403).send('Unauthorized');
+        }
     },
 
     deleteBook: (req, res) => {
-        const db = req.app.get('db');
-        const { bookId } = req.params;
+        if(req.user.is_admin){
 
-        db.delete_book([bookId]).then(deleted => {
-            res.status(200).send('Book deleted successfully')
-        })
+            const db = req.app.get('db');
+            const { bookId } = req.params;
+            
+            db.delete_book([bookId]).then(deleted => {
+                res.status(200).send('Book deleted successfully')
+            })
+        } else {
+            res.status(403).send('Unauthorized')
+        }
     }
 
 }
